// resources/js/types/lead.d.ts
export interface Lead {
    id: number;
    first_name: string;
    last_name: string;
    email: string;
    mobile_country_code: string;
    mobile: string;
    is_whatsapp: boolean;
    country_of_residence: string;
    in_uk_now: boolean;
    study_type: string;
    subject_interested: string;
    created_at: string;
    updated_at: string;
}